DB周りの修正 organization用DBへの接続はITA共通DBに取得しにいく
